requirements.txt
discord.py
datetime